Tightening dependencies and cfg

+++ workspace/swiss-army-knife/src/simd_compatibility/x86_64/_mm_extract_epi8_wrapper.rs
@@ -10,9 +10,9 @@ pub unsafe fn _mm_extract_epi8_wrapper<const imm8: i32>(a: std::arch::x86_64::__
 	std::arch::x86_64::_mm_extract_epi8(a, imm8)
 }
 
+/// Wrapper function for `_mm_extract_epi8` to support constant and non-constant forms in Rust and our compatibility module.
 #[cfg(all(target_feature = "sse2", not(target_feature = "sse4.1")))]
-#[target_feature(enable = "sse4.1")]
 pub unsafe fn _mm_extract_epi8_wrapper<const imm8: i32>(a: std::arch::x86_64::__m128i) -> i32
 {
-	sse2::_mm_extract_epi8_constant::<imm8>(a)
+	sse2::_mm_extract_epi8_constant::<imm8>(a) as i32
 }
